Add team onboarding section
This is a cleaned up version of the internal
onboarding document
(https://docs.google.com/document/d/1bJSrlyc-e7bcOxV4sjx3FesMNVgdwNxUzMvIYywbt-0)
with the intention to replace it. It still has many
links to internal documents but is thought to serve
as the top level starting point for people onboarding
to Red Hat OpenShift Monitoring teams.

Over time more and more of the internal links can be
moved to this repository as well.